\name{HIV.LifeTables-internal}

\alias{Mx.svd.scores.m}
\alias{Mx.svd.scores.f}
\alias{svd.coeffs.m}
\alias{svd.coeffs.f}
\alias{svd.coeffs.xp.m}
\alias{svd.coeffs.xp.f}
\alias{co1.e0mod.a.m}
\alias{co1.e0mod.a.f}
\alias{co1.e0mod.na.m}
\alias{co1.e0mod.na.f}
\alias{co1.5q0mod.a.m}
\alias{co1.5q0mod.a.f}
\alias{co1.5q0mod.na.m}
\alias{co1.5q0mod.na.f}
\alias{co1.45q15mod.a.m}
\alias{co1.45q15mod.a.f}
\alias{co1.45q15mod.na.m}
\alias{co1.45q15mod.na.f}
\alias{co2.e0mod.a.m}
\alias{co2.e0mod2.a.m}
\alias{co2.e0mod3.a.m}
\alias{co2.e0mod.a.f}
\alias{co2.e0mod2.a.f}
\alias{co2.e0mod3.a.f}
\alias{co2.e0mod.na.m}
\alias{co2.e0mod2.na.m}
\alias{co2.e0mod3.na.m}
\alias{co2.e0mod.na.f}
\alias{co2.e0mod2.na.f}
\alias{co2.e0mod3.na.f}
\alias{co2.5q0mod.a.m}
\alias{co2.5q0mod2.a.m}
\alias{co2.5q0mod3.a.m}
\alias{co2.5q0mod.a.f}
\alias{co2.5q0mod2.a.f}
\alias{co2.5q0mod3.a.f}
\alias{co2.5q0mod.na.m}
\alias{co2.5q0mod2.na.m}
\alias{co2.5q0mod3.na.m}
\alias{co2.5q0mod.na.f}
\alias{co2.5q0mod2.na.f}
\alias{co2.5q0mod3.na.f}
\alias{co2.45q15mod.a.m}
\alias{co2.45q15mod2.a.m}
\alias{co2.45q15mod3.a.m}
\alias{co2.45q15mod.a.f}
\alias{co2.45q15mod2.a.f}
\alias{co2.45q15mod3.a.f}
\alias{co2.45q15mod.na.m}
\alias{co2.45q15mod2.na.m}
\alias{co2.45q15mod3.na.m}
\alias{co2.45q15mod.na.f}
\alias{co2.45q15mod2.na.f}
\alias{co2.45q15mod3.na.f}
\alias{co3.e0mod.a.m}
\alias{co3.e0mod2.a.m}
\alias{co3.e0mod3.a.m}
\alias{co3.e0mod.a.f}
\alias{co3.e0mod2.a.f}
\alias{co3.e0mod3.a.f}
\alias{co3.e0mod.na.m}
\alias{co3.e0mod2.na.m}
\alias{co3.e0mod3.na.m}
\alias{co3.e0mod.na.f}
\alias{co3.e0mod2.na.f}
\alias{co3.e0mod3.na.f}
\alias{co3.5q0mod.a.m}
\alias{co3.5q0mod2.a.m}
\alias{co3.5q0mod3.a.m}
\alias{co3.5q0mod.a.f}
\alias{co3.5q0mod2.a.f}
\alias{co3.5q0mod3.a.f}
\alias{co3.5q0mod.na.m}
\alias{co3.5q0mod2.na.m}
\alias{co3.5q0mod3.na.m}
\alias{co3.5q0mod.na.f}
\alias{co3.5q0mod2.na.f}
\alias{co3.5q0mod3.na.f}
\alias{co3.45q15mod.a.m}
\alias{co3.45q15mod2.a.m}
\alias{co3.45q15mod3.a.m}
\alias{co3.45q15mod.a.f}
\alias{co3.45q15mod2.a.f}
\alias{co3.45q15mod3.a.f}
\alias{co3.45q15mod.na.m}
\alias{co3.45q15mod2.na.m}
\alias{co3.45q15mod3.na.m}
\alias{co3.45q15mod.na.f}
\alias{co3.45q15mod2.na.f}
\alias{co3.45q15mod3.na.f}
\alias{ages}
\alias{user.par}
\alias{africa.nums}
\alias{la.nums}
\alias{mu.age.f}
\alias{mu.age.m}
\alias{sd.age.f}
\alias{sd.age.m}

\title{
Internal models and data sets of HIV.LifeTables
}

\description{
These functions and data sets are not intended be used directly by the user but the coefficients for modeling the weights used in this system can be obtained from the object titled "co1.e0mod.a.m" where the first portion of the name refers to the weight (1-3), the second part refers to the model (e0mod, 5q0mod, or 45q15mod)*, the third portion of the name refers to the region (a for Africa, na for Caribbean), and the final portion indicates sex (m for males, f for female).

*If a '2' or '3' is added to this is portion of the model name, '2' indicates adult ART is included as a predictor and '3' indicates both child and adult ART are included as predictors.
}

\keyword{internal}