\name{HIV.LifeTables-internal}

\alias{Mx.svd.scores}
\alias{svd.coeffs}
\alias{svd.coeffs.xp}
\alias{co1.e0mod.a.m}
\alias{co1.e0mod.a.f}
\alias{co1.e0mod.na.m}
\alias{co1.e0mod.na.f}
\alias{co1.5q0mod.a.m}
\alias{co1.5q0mod.a.f}
\alias{co1.5q0mod.na.m}
\alias{co1.5q0mod.na.f}
\alias{co1.45q15mod.a.m}
\alias{co1.45q15mod.a.f}
\alias{co1.45q15mod.na.m}
\alias{co1.45q15mod.na.f}
\alias{co2.e0mod.a.m}
\alias{co2.e0mod.a.f}
\alias{co2.e0mod.na.m}
\alias{co2.e0mod.na.f}
\alias{co2.5q0mod.a.m}
\alias{co2.5q0mod.a.f}
\alias{co2.5q0mod.na.m}
\alias{co2.5q0mod.na.f}
\alias{co2.45q15mod.a.m}
\alias{co2.45q15mod.a.f}
\alias{co2.45q15mod.na.m}
\alias{co2.45q15mod.na.f}
\alias{co3.e0mod.a.m}
\alias{co3.e0mod.a.f}
\alias{co3.e0mod.na.m}
\alias{co3.e0mod.na.f}
\alias{co3.5q0mod.a.m}
\alias{co3.5q0mod.a.f}
\alias{co3.5q0mod.na.m}
\alias{co3.5q0mod.na.f}
\alias{co3.45q15mod.a.m}
\alias{co3.45q15mod.a.f}
\alias{co3.45q15mod.na.m}
\alias{co3.45q15mod.na.f}
\alias{africa.nums}
\alias{la.nums}

\title{
Internal models and datasets of LifeTables
}

\description{
These functions and datasets are not intended be used directly by the user but the coefficients for modeling the weights used in this system can be obtained from the object titled "co1.e0mod.a.m" where the first portion of the name refers to the weight (1-3), the second part refers to the model (e0mod, 5q0mod, or 45q14mod), the third portion of the name refers to the region (a for Africa, na for Caribbean), and the final portion indicates sex (m for males, f for female).
}

\keyword{internal}